app/requirements.txt
streamlit==0.89.0
pandas==1.3.3
numpy==1.21.2
psycopg2-binary==2.9.1
